1. changed getAngle to return float in order to make it consistent with other methods taking angle as a float parameter 2. made all variables protected (some were private, others protected and some public) 3. split setSpeed in a speed for moving (e.g., cm/s) and one for turning (°/s) 4. provided methods to get maximum movement and turning speed (if you don't know the physics the robot needs to tell you its max) 5. depreciated setSpeed(int), as this method requires knowledge about the physics of your robot (which you should not need) 6. replaced drift correction by allowing two different wheel sizes, as it is the more accurate and general approach

+++ trunk/classes/lejos/navigation/CompassPilot.java
@@ -101,8 +101,8 @@ public class CompassPilot extends TachoPilot {
 	/**
 	 * Returns the compass angle in degrees, Cartesian (increasing counter clockwise) i.e. the actual robot heading
 	 */
-	public int getAngle() {
-		return (int)compass.getDegreesCartesian();
+	public float getAngle() {
+		return compass.getDegreesCartesian();
 	}
 	
 	/**
@@ -134,7 +134,7 @@ public class CompassPilot extends TachoPilot {
 	 */
 	private int getHeadingError() 
 	{
-	   int  err = getAngle() - _heading;
+	   int  err = (int)getAngle() - _heading;
 		// Handles the wrap-around problem:
 		while (err < -180) err = err + 360;
 		while (err > 180) err = err - 360;
